requirements.txt
legacy~=0.1.6
pip~=22.3
argparse~=1.4.0
Jinja2~=3.1.2
future~=0.18.2
pytz~=2022.6
MarkupSafe~=2.1.1
Werkzeug~=2.2.2
click~=8.1.3
itsdangerous~=2.1.2
Django~=4.1.3
asgiref~=3.5.2
sqlparse~=0.4.3
colorama~=0.4.6
requests~=2.28.1
gevent~=22.10.2
cffi~=1.15.1
greenlet~=2.0.1
automium~=0.2.6
Flask~=2.2.2
certifi~=2022.9.24
six~=1.16.0
idna~=3.4
urllib3~=1.26.13
pycparser~=2.21
tldextract~=3.4.0
setuptools~=65.5.0
filelock~=3.8.1
djangorestframework~=3.14.0